1、系统框架结构调整； 2、核销相关功能完善； 3、收银功能完善；

+++ fuint-application/application-service/src/main/java/com/fuint/application/service/confirmlog/ConfirmLogService.java
@@ -27,7 +27,7 @@ public interface ConfirmLogService {
      * @param userCouponId
      * @return
      * */
-    Integer getConfirmNum(Integer userCouponId) throws BusinessCheckException;
+    Long getConfirmNum(Integer userCouponId) throws BusinessCheckException;
 
     /**
      * 获取核销总数
